requirements.txt
clueai==0.0.2.2.4
fake_useragent==1.1.3
Flask==2.2.2
psutil==5.9.0
PyYAML==6.0
redis==4.5.4
Requests==2.29.0
selenium==3.3.1

gevent~=22.10.2
colorama~=0.4.6